Add defines to make the compilation works for both desktop and web versions. Running just 'make' compiles the code for desktop and running 'make PLATFORM=PLATFORM_WEB -B' compiles the code for web. The qif library must already be compiled before qif-graphics according to the desired version, web or desktop.

+++ src/data.cpp
@@ -290,7 +290,7 @@ void Data::newRandomChannel(int num_out){
 
         random_shuffle(prob.begin(), prob.end());
 
-        for(long unsigned int j = 0; j < num_out; j++){
+        for(int j = 0; j < num_out; j++){
             channel[i][j] = prob[j];
         }
     }
